+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 Version 2.1.1 (in development)
 ------------------------------
 
+* exposed adjacency and incidence (see issues #33, #34, and #36,
+  contributed by bobmyhill)
+
 Version 2.1.0 (15 October 2018)
 -------------------------------